separate Actions and Interfaces
