+++ src/Interfaces/Animations.ts
