+++ src/Interfaces/GraphicsEngine.ts
