+++ src/Interfaces/InteractionEvents/MemoEvents.ts
