+++ src/Interfaces/Memos.ts
