+++ src/Interfaces/Snapshot.ts
